requirements.txt
aiofile
aiopg
asyncpg
databases[postgresql]
fastapi
httpx
psycopg2
python-jose[cryptography]
pyyaml
requests
SQLAlchemy
uvicorn